+++ src/Tortoise.Windows/Extensions/ServiceCollectionExtensions.cs
@@ -1,7 +1,9 @@
 using System.Runtime.Versioning;
 using Microsoft.Extensions.DependencyInjection;
 using Tortoise.Core.Devices;
+using Tortoise.Core.Drivers;
 using Tortoise.Windows.Devices;
+using Tortoise.Windows.Drivers;
 
 namespace Tortoise.Windows.Extensions;
 
@@ -11,6 +13,15 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddTortoiseWindowsDeviceInventory(this IServiceCollection services)
     {
         services.AddSingleton<IDeviceInventoryProvider, WindowsDeviceInventoryProvider>();
+        return services;
+    }
+
+    [SupportedOSPlatform("windows")]
+    public static IServiceCollection AddTortoiseWindowsDriverPackageInventory(this IServiceCollection services)
+    {
+        services.AddTortoiseWindowsDeviceInventory();
+        services.AddSingleton<IDriverPackageInventoryProvider, WindowsDriverPackageInventoryProvider>();
+        services.AddSingleton<IDriverPackageExportService, DisabledDriverPackageExportService>();
         return services;
     }
 }